requirements.txt
monkey-patch.py
pydantic==2.4.2
openai~=0.28.1
python-dotenv~=1.0.0